backend/requirements.txt
Flask==3.0.3
itsdangerous==2.2.0
mysql-connector-python==8.0.33
icecream==2.1.3